refractor : short[][][] to int[][]
