+++ java/module-screen-video/src/main/java/com/swe/ScreenNVideo/Synchronizer/ImageSynchronizer.java
@@ -1,0 +1,18 @@
+package com.swe.ScreenNVideo.Synchronizer;
+
+import com.swe.ScreenNVideo.PatchGenerator.CompressedPatch;
+
+import java.util.List;
+
+public class ImageSynchronizer {
+    int[][] previousImage;
+
+    public ImageSynchronizer() {
+        previousImage = null;
+    }
+
+    public int[][] synchronize(List<CompressedPatch> compressedPatches) {
+
+    }
+
+}
